test: cast values to fix WIN32 warnings

+++ src/test.c
@@ -535,7 +535,7 @@ static int testcase_perf(const struct test *test, double *usec_avgp)
 
 	usec_avg = 1.0 * (usec_stop - usec_start) / (double)i;
 
-	n = usec_avg ? (REPEATS_USEC / usec_avg) : 0;
+	n = usec_avg ? (size_t)(REPEATS_USEC / usec_avg) : 0;
 	n = min(REPEATS_MAX, max(n, REPEATS_MIN));
 
 	/* now for the real measurement */
@@ -647,7 +647,7 @@ int test_perf(const char *name, bool verbose)
 				return err;
 			}
 
-			tim->nsec_avg = 1000.0 * usec_avg;
+			tim->nsec_avg = (uint64_t)(1000.0 * usec_avg);
 		}
 
 		/* sort the timing table by average time */
